Organized Customer-Service directories
List of changes:

- Removed OwnerMapper
- Remove ServiceUtil
- Removed Duplicate Database Files
- Removed Unused Import
- Removed Mapstruct from build.gradle
- Renamed utils to CustomerExceptions
