+++ customers-service/build.gradle
@@ -10,10 +10,6 @@ group = 'com.petclinic.customers'
 version = '1.0.0-SNAPSHOT'
 sourceCompatibility = '1.8'
 
-ext {
-	mapstructVersion = "1.3.1.Final"
-}
-
 repositories {
 	mavenCentral()
 }
@@ -24,11 +20,6 @@ dependencies {
 	implementation('org.springframework.boot:spring-boot-starter-data-jpa')
 	implementation('mysql:mysql-connector-java')
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
-	implementation("org.mapstruct:mapstruct:${mapstructVersion}")
-
-	compileOnly "org.mapstruct:mapstruct-processor:${mapstructVersion}"
-	annotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
-	testAnnotationProcessor "org.mapstruct:mapstruct-processor:${mapstructVersion}"
 
 	runtimeOnly 'com.h2database:h2'
