--- customers-service/src/main/java/com/petclinic/customers/businesslayer/OwnerMapper.java
@@ -1,26 +1,0 @@
-package com.petclinic.customers.businesslayer;
-
-/*
-import com.petclinic.customers.datalayer.Owner;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import com.petclinic.bffapigateway.dtos.*;
-import org.mapstruct.Mappings;
-
-*/
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-
-public interface OwnerMapper {
-
-    /*
-    @Mapping(target = "id",ignore = true)
-    OwnerDetails entityToModel(Owner entity);
-
-    @Mappings({
-            @Mapping(target = "")
-    })
-   */
-
-}
