--- customers-service/src/main/resources/schema-mysql.sql
@@ -1,28 +1,0 @@
-USE `customers-db`;
-
-CREATE TABLE IF NOT EXISTS types (
-    id INT(4) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    name VARCHAR(80),
-    INDEX(name)
-    ) engine=InnoDB;
-
-CREATE TABLE IF NOT EXISTS owners (
-    id INT(4) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    first_name VARCHAR(30),
-    last_name VARCHAR(30),
-    address VARCHAR(255),
-    city VARCHAR(80),
-    telephone VARCHAR(20),
-    INDEX(last_name)
-    ) engine=InnoDB;
-
-CREATE TABLE IF NOT EXISTS pets (
-    id INT(4) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    name VARCHAR(30),
-    birth_date DATE,
-    type_id INT(4) UNSIGNED NOT NULL,
-    owner_id INT(4) UNSIGNED NOT NULL,
-    INDEX(name),
-    FOREIGN KEY (owner_id) REFERENCES owners(id),
-    FOREIGN KEY (type_id) REFERENCES types(id)
-    ) engine=InnoDB;
